rust: helpers: Remove uaccess helpers
Now that we support wrap-static-fns we no longer need the custom helper.

Tested-by: Dirk Behme <dirk.behme@de.bosch.com>

+++ rust/kernel/uaccess.rs
@@ -226,7 +226,8 @@ impl UserSliceReader {
         }
         // SAFETY: `out_ptr` points into a mutable slice of length `len`, so we may write
         // that many bytes to it.
-        let res = unsafe { bindings::copy_from_user(out_ptr, self.ptr as *const c_void, len) };
+        let res =
+            unsafe { bindings::copy_from_user(out_ptr, self.ptr as *const c_void, len as u64) };
         if res != 0 {
             return Err(EFAULT);
         }
@@ -330,7 +331,7 @@ impl UserSliceWriter {
         }
         // SAFETY: `data_ptr` points into an immutable slice of length `len`, so we may read
         // that many bytes from it.
-        let res = unsafe { bindings::copy_to_user(self.ptr as *mut c_void, data_ptr, len) };
+        let res = unsafe { bindings::copy_to_user(self.ptr as *mut c_void, data_ptr, len as u64) };
         if res != 0 {
             return Err(EFAULT);
         }
